Moved the NumpyEncoder class from the parameter sweep file to the stochss encoders file to avoid circular imports.
